module5/workoutgateway/requirements.txt
Flask==3.0.0
opentelemetry-distro==0.41b0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
requests==2.31.0
flask-cors==4.0.0
